+++ R/assertions.R
@@ -1,10 +1,36 @@
 # is npm installed and can it be found?
-has_npm <- function(){
-  length(npm_find()) > 0
+has_engine <- function(){
+  length(engine_find()) > 0
 }
 
-assertthat::on_failure(has_npm) <- function(call, env){
-  stop("Cannot find `npm`, are you sure it is installed?", call. = FALSE)
+assertthat::on_failure(has_engine) <- function(call, env){
+  stop(
+    sprintf("Cannot find engine `%s`, are you sure it is installed?", engine_get()), 
+    call. = FALSE
+  )
+}
+
+# is npm installed and can it be found?
+is_yarn <- function(){
+  engine_get() == "yarn"
+}
+
+assertthat::on_failure(is_yarn) <- function(call, env){
+  stop(
+    "yarn is not your current engine, see `engine_set()`.", 
+    call. = FALSE
+  )
+}
+
+is_npm <- function(){
+  engine_get() == "npm"
+}
+
+assertthat::on_failure(is_npm) <- function(call, env){
+  stop(
+    "npm is not your current engine, see `engine_set()`.",
+    call. = FALSE
+  )
 }
 
 # check that argument is not missing