+++ R/loaders.R
@@ -3,19 +3,33 @@
 #' Installs loaders and adds relevant configuration rules to `srcjs/config/loaders.json`.
 #' 
 #' @param test Test regular expression test which files should be transformed by the loader.
+#' @param import Whether to enable `import` statements for 
+#' `.css` files. If `FALSE` use `require`.
+#' @param modules Enables CSS modules and their config,
+#' a complex but powerful feature detailed 
+#' [here](https://webpack.js.org/loaders/css-loader/#modules)
 #' 
 #' @details This will let you import styles much like any other modules, e.g.: `import './styles.css'`.
 #' 
 #' @section Packages:
 #' 
-#' * [use_loader_css()] - installs and imports `style-loader` and `css-loader` packages as dev.
+#' * [use_loader_css()] - installs and imports `css-loader` packages as dev.
+#' * [use_loader_style()] - installs and imports `style-loader` and `css-loader` packages as dev. This loader enabled CSS modules.
 #' * [use_loader_sass()] - installs and imports `style-loader`, `css-loader`, and `sass-loader` as dev.
 #' 
 #' @name style_loaders
 #' @export
-use_loader_css <- function(test = "\\.css$"){
+use_loader_css <- function(test = "\\.css$", import = TRUE, modules = TRUE){
   assert_that(has_scaffold())
-  use_loader_rule(c("style-loader", "css-loader"), test = test)
+  use_loader_rule(
+    "css-loader", 
+    test = test, 
+    options = list(
+      import = TRUE,
+      modules = TRUE
+    ),
+    .name_use = "loader" # allows options
+  )
 }
 
 #' @rdname style_loaders
@@ -23,6 +37,25 @@ use_loader_css <- function(test = "\\.css$"){
 use_loader_sass <- function(test = "\\.s[ac]ss$/i"){
   assert_that(has_scaffold())
   use_loader_rule(c("style-loader", "css-loader", "sass-loader"), test = test)
+}
+
+#' @rdname style_loaders
+#' @export
+use_loader_style <- function(test = "\\.css$", import = TRUE, modules = TRUE){
+  assert_that(has_scaffold())
+  use_loader_rule(
+    c("css-loader", "style-loader"), 
+    test = test, 
+    use = list(
+      "style-loader",
+      list(
+        loader = "css-loader",
+        options = list(
+          modules = TRUE
+        )
+      )
+    )
+  )
 }
 
 #' Use Pug Loader
@@ -35,6 +68,24 @@ use_loader_sass <- function(test = "\\.s[ac]ss$/i"){
 use_loader_pug <- function(test = "\\.pug$"){
   assert_that(has_scaffold())
   use_loader_rule("pug-loader", test = test)
+}
+
+#' Use Typescript Loader
+#' 
+#' Adds the loader for the pug templating engine.
+#' 
+#' @inheritParams style_loaders
+#' 
+#' @export 
+use_loader_ts <- function(test = "\\.tsx?$"){
+  assert_that(has_scaffold())
+  use_loader_rule(
+    c("typescript", "ts-loader"), 
+    test = test, 
+    use = "ts-loader",
+    exclude = "/node_modules/"
+  )
+  create_ts_config()
 }
 
 #' Use babel Loader
@@ -54,7 +105,7 @@ use_loader_pug <- function(test = "\\.pug$"){
 use_loader_babel <- function(test = "\\.(js|jsx)$", use_eslint = FALSE){
   assert_that(has_scaffold())
 
-  pkgs <- "babel-loader"
+  pkgs <- c("babel-loader", "@babel/core")
   if(use_eslint) pkgs <- c(pkgs, "eslint-loader")
   use_loader_rule(pkgs, test = test, exclude = "/node_modules/")
 }
@@ -132,6 +183,11 @@ use_loader_file <- function(test = "\\.(png|jpe?g|gif)$/i"){
 #' @export 
 use_loader_eslint <- function(test = "\\.(js|jsx)$"){
   assert_that(has_scaffold())
+  .Deprecated(
+    "add_plugin_eslint",
+    package = "packer",
+    "The loader will soon be deprecated, it is advised to use the plugin instead."
+  )
   use_loader_rule("eslint-loader", test)
 }
 
@@ -142,16 +198,18 @@ use_loader_eslint <- function(test = "\\.(js|jsx)$"){
 #' @inheritParams style_loaders
 #' @param packages NPM packages (loaders) to install.
 #' @param use Name of the loaders to use for `test`.
+#' @param .name_use Dependending on the webpack config 
+#' one might want to change the `use` to `loader` or `loaders`.
 #' @param ... Any other options to pass to the rule.
 #' 
 #' @details Reads the `srcsjs/config/loaders.json` and appends the rule.
 #' 
 #' @export 
-use_loader_rule <- function(packages, test, ..., use = as.list(packages)){
+use_loader_rule <- function(packages, test, ..., use = as.list(packages), .name_use = "use"){
   assert_that(has_scaffold())
   assert_that(not_missing(packages))
 
-  npm_install(packages, scope = "dev")
+  engine_install(packages, scope = "dev")
 
   # message modifications
   loader <- list(
@@ -159,9 +217,14 @@ use_loader_rule <- function(packages, test, ..., use = as.list(packages)){
     use = use,
     ...
   )
+
+  # rename use to allow options
+  names(loader)[2] <- .name_use
+
+  # add the loader
   loader_add(loader)
   
-  # wrap up
+  # message
   loader_msg(packages)
 }
 
@@ -199,4 +262,55 @@ loader_add <- function(loader){
 #' @keywords internal
 loader_msg <- function(loaders){
   cli::cli_alert_success("Added loader rule for {.val {loaders}}\n")
+}
+
+#' Basic TypeScript Config
+#' @keywords internal
+#' @noRd 
+create_ts_config <- function(){
+
+  if(file.exists("tsconfig.json")){
+    cli::cli_alert_info("{.file tsconfig.json} already exists")
+    return()
+  }
+
+  out <- tryCatch(
+    jsonlite::fromJSON("srcjs/config/output_path.json"),
+    error = "./inst/packer"
+  )
+  
+  replace_entry_point()
+
+  config <- list(
+    compilerOptions = list(
+      outDir = out[1], 
+      noImplicitAny = TRUE, 
+      module = "es6", 
+      target = "es5"
+    )
+  )
+
+  jsonlite::write_json(config, "tsconfig.json", pretty = TRUE, auto_unbox = TRUE)
+  cli::cli_alert_success("Writing {.file tsconfig.json}")
+  usethis::use_build_ignore("tsconfig.json")
+}
+
+replace_entry_point <- function(){
+
+  infile <- tryCatch(
+    jsonlite::fromJSON("srcjs/config/entry_points.json"),
+    error = NULL
+  )
+
+  if(is.null(infile))
+    return()
+
+  nms <- names(infile)
+  infile <- gsub("\\.js", "\\.ts", infile)
+  names(infile) <- nms
+
+  infile <- as.list(infile)
+
+  jsonlite::write_json(infile, "srcjs/config/entry_points.json", pretty = TRUE, auto_unbox = TRUE)
+
 }